Move export into own file
